Multiplayer works!
The pieces sync states with each other (although basically, this
will be improved in the future), and only one player can grab
pieces at a time. Any player, including the dedicated server, can
add pieces to the table.

This is following the model that a singleplayer game will be a
multiplayer game with a maximum of 1 player, to make the scripting
easier and nicer.