Started working on the Stack piece.
Although, I am having trouble with its physics once it stores more
than two pieces. It seems to randomly ignore gravity once the third
item is put on, until the mass is changed manually in the inspector.

Plus, putting the stack on top of a singular piece does not work yet,
as the stack does not detect a collision from the piece, only the table.