Added the StackPieceInstance.
Instead of moving stackable pieces to being children of the stack
directly, instead, the mesh and collision shape of the piece is
extracted beforehand, and the stack has the mesh instance as the child
instead. This fixes the previous problem of the stack sometimes not
behaving correctly when hovering, and putting a stack on top of the
stackable piece now works.