Optimised sending the room state to new clients.
Instead of sending a rpc to the new client for each piece, instead,
the game will gather the room state into a dictionary, and will send
the dictionary in one go.

Also added enums to the Stack class, which allows the room to either
manually set whether a piece goes on top or bottom, or is flipped or
not, or allow the stack to set these automatically.

Also removed the y-axis from determining whether two stackable pieces
have stacked, as a stack can grow quite tall, but its origin stays at
the bottom of the stack, thus if it was placed on a piece upside down,
it would eventually break the threshold.