Added the ShuffleableStack.
When two cards stack, by default, it creates a ShuffleableStack,
which shuffles the piece order upon shaking.

Also modified the Piece.is_being_shaked() function to only return
true if the new velocity is in the opposite direction to the old
velocity, as pieces were being shaked when they we only being moved
around the table really fast.