Prefixed vars and funcs with srv_ if they are only used by the server.
For variables: The prefix should apply if the variable is only useful
to the server, and reading it client-side won't make sense.

For functions: The prefix should apply if the function should only be
run on the server-side (i.e. if it uses a server-only variable), AND
it hasn't already been declared a "master" remote function.