Moved the piece scale information to configuration files.
The piece scenes in the game project have been made to have a
unit default, either a unit cube or a unit sphere, depending on the
shape.

This way, users can make standard pieces any size they want, instead
of being stuck with a particular size for a given type of piece.