Added a context menu when pieces are right-clicked.
Right now, the menus only have one button, "Delete", to remove the
piece from the room. However, this can be easily modified to add
other controls to the context menu by changing the
'add_context_to_control' function of the piece.

However, if one were to override that function, they must remember
to call the parents version of the function as well with
'.add_context_to_control()'!