Added the ability to start the game as a dedicated server.
If the game is running in Release mode, if the server.cfg file exists,
then the game will automatically start the dedicated server.

If the game is running in Debug mode, if the server.cfg file exists,
then the game will show a "Start Dedicated Server" button on the main
menu.