Added the ability to save and load table states.
With the new buttons added to the in-game menu, the user can save
the current table state to a .table file, and load it again.

In multiplayer, currently all clients can tell the server to load
any given state, which should be changed once permissions and hosts
have been implemented.

This fixes #10.