Can now flip the table! :D
Once the table is flipped, it can then be reset to the state it was
in just before it was flipped.

When saving the game post-flip, the state of the room keeps track of
the pre-flip state, so when you load the game again, you can still
reset the table.

This fixes #17.