Added a mouse grabbing control to the camera controller.
All of the code that handles mouse input events has been moved to
a new function, which is connected to this new control's gui_input
signal.

This is so it can grab focus from the chat box, which previously
would not let go of it's focus when clicking away from it.