Added the chess pieces and board to the default asset pack.
However, one thing I forgot to account for when exporting the
pieces is that the bottom of the pieces were at y = 0, meaning that
when they are imported into Godot, the centre of mass is at the
bottom of the piece, which is a bug.

On the other hand, I don't think the user should have to remember
to translate their pieces to match the centre of mass of the object.
So, I will try to make it such that the game will try and correct
the centre of mass as best as it can before spawning the piece in.