Bounding box cache is no longer stored scaled.
This meant that even if the scale of the piece was changed in the
config file, it would not update the bounding box of the piece
(even though in reality it had changed), since the cache did not
update.

Now, the cache stores the bounding box pre-scale, so the scale will
always be applied to the bounding box, even if it came from the
cache.