Added a measuring tool.
Two extra buttons have been added to the camera controller: One for
selecting the cursor tool (the "original" functionality), and the
other is for selecting the new measuring tool.

The measuring tool will measure the distance between two points made
by the player (by left-clicking with the tool selected). The tool
will show the distance in both cm and in. However, the player cannot
move or right-click any pieces on the table while this tool is
selected.

This commit fixes #6.