Can now import audio files which are played in the main menu.
Audio files can now be imported via the music/ and sounds/ folders
in asset packs. If configured to, the music has a chance of being
played in the main menu.

I also included some jazz tracks from the one and only Kevin
MacLeod, which serve two purposes: to help test the upcoming
features, and to set the "mood" I want the game to have, e.g. the
game is cozy, chill, relaxed.