Added the speaker object, which right now just plays one track.
The audio is positional, meaning the audio changes depending on
where the speaker is relative to the camera.

The idea is that players will be able to right-click the speaker
and choose what track they want to play, and this will be synced
for all players in a multiplayer game.

Also added a compressor to the master audio bus, because it turns
out if you spawn more than 100 of these speakers, without a
compressor the audio gets a bit loud :O