Fixed bugs caused by the factory when the search bar was used.
One major bug was that the game hanged if the search bar was used
while the factory was building pieces. This happened because the
preview grid flushed the jobs before setting the previews, which
(somehow) hanged the create_convex_shape function, which in turn
hanged the main thread.

This was fixed by having the build thread de-allocate itself once
it was done.

Another bug was that if the seach bar was used frequently, the
factory was setting preview displays after the search term was
changed.

This was fixed by having the previews reject incoming factory
pieces if they recently had their details cleared.