The room now loads the table from the default asset pack.
However, there is a slight issue to do with collisions now that the
table's collision shape is automatically generated - small pieces
don't fall through the table, but they jitter a lot before they
"settle".

Setting the physics FPS to be the refresh rate of my monitor fixes
the issue, but this is obviously not a complete fix - I'm going to
try and do some more research and ask around to try and find out
what the root problem is.