Camera movement is now updated every frame.
The movement code was being called during the physics frame,
meaning that players with refresh rates higher than 60Hz were able
to tell that the movement was not as smooth as their monitors were
allowing it to be.

However, this has also revealed a flaw in the game for players with
refresh rates higher than the physics FPS (60Hz) - the physics
engine does not interpolate the movement of objects in-between
physics frames, so objects appear "laggy".