Cards are now added to hands upon loading a state.
States do not store information about what cards are in what hands
(because hands can differ between play sessions), instead, when the
state is loaded, the room waits a few physics frames for the hands
to detect the cards, then it will assign the cards to the hand they
are in.

This commit fixes #25.