Added a progress bar to the importing assets scene.
When importing assets, the AssetDB now first generates a catalog of
the files it will import (this is used to calculate the total number
of files it will import), then it will go ahead with importing them
as usual.

Stacks are now also stored in the same type as the objects they
contain (e.g. "cards", "tokens/cylinder") instead of in their own
type ("stacks").