Added a paint and erase tool for painting on the table.
Players can now paint on the table with any colour, and with
varying brush sizes. They can also erase paint on the table with
the erase tool.

The paint plane uses a custom shader to put paint on the table, and
the next frame the new viewport image is saved, and the cycle repeats
as long as the player is painting/erasing.

Since the paint plane viewport is quite small (512 x 512), bicubic
filtering is used to smooth the image when displayed on the table.

The size of the paint plane can be configured via the asset pack's
tables/config.cfg file.

This commit fixes #35.