Added the StackSandwich class.
This is an optimised version of the Stack class for stacks of cards.
Only the top face of the top card and the bottom face of the bottom
card are rendered, and the sides are based on the edges of one of
the faces.

This means that potentially thousands of cards could be in a stack
with little performance deficit.