Game now creates a TabletopClub folder in the Documents folder.
At the start of the game, it will create this folder, and place an
"assets" subfolder inside if it doesn't already exist, to provide
the user a convenient, easy-to-find place to place custom asset
packs.

The game also creates a "saves" subfolder once in-game, which is now
the default location for manually-created save files. This will also
be the location of future automatic save files.

In the future screenshots will also probably be put in a subfolder
here as well.