Added autosaves that are created at a configurable interval.
The ability to change the interval, and how many autosaves can be
created, has been added to the options menu.

However, autosaves take up to 100ms to create, which is noticeable
in-game. The main culprit seems to be the room's get_state() function,
so I will try to optimise this function at some point.