Reduced the network traffic from client to server when hovering.
Previously, when the client hovered multiple pieces, it sent the
updated hover position for every piece, even though the hover
position was the same for all of them, which caused redundant packets.

Now, when the initial hover request is made for multiple pieces, the
server stores the names of the pieces for future reference, so that when
the client sends an updated hover position, it can update the hover
position for all of the hovering pieces, without the client having to
send multiple requests. Now, the network traffic from the client to the
server is constant, and independent of the number of pieces being hovered.

This also revealed another potential optimisation: the server relies on
the automatic state update method to send the new positions to the other
clients. Since it does this for all moving pieces every physics tick, this
could cause quite a bit of network traffic (upwards of 500MiB/s) if lots of
pieces are being hovered. Therefore, when pieces are being hovered, we could
instead just send one hover position to all clients, provided all of the
clients know which pieces are being hovered.

This optimisation will be done in the future, as well as a possible
re-write of the Piece class.