Re-written the Piece class such that all clients hover the piece.
This is the start of a series of commits to optimise the network
traffic when hovering pieces. These changes allow all clients to
know who is hovering a piece, and the server no longer sends state
updates when a piece is being hovered - it is now up to the clients
to apply the custom forces when it receives the new hover position
and basis from the server.