Changed the rendering thread model to multi-threaded.
This is recommended by the Godot docs, since we instance nodes that
render objects in 3D in separate threads.

This *should* stop a bug I've been noticing in the main menu where
it sometimes hangs when loading into either singleplayer or multiplayer.
However, I cannot confirm the fix 100%, I can only confirm that the bug
has not appeared since.