Fixed index out of bounds error in the notebook when moving pages.
The issue was caused by the fact that the Move Up and Move Down
buttons were not enabled/disabled when pressed, so their state stayed
the same, even if the player reached the top or bottom of the list.
This caused the _last_page_seen index to become invalid.

This commit fixes #83.