Added the flick tool.
This tool allows the player to apply an impulse to a piece from
a given position. In other words, it makes pieces go flying :D

There is also an option to modify the strength of the flick, so that
players can flick heavier objects, or send smaller objects into orbit.