Added a limbo system for removing pieces client-side.
For clients, when a piece is removed, instead of removing them
directly from the scene tree, they stay, but in a "limbo" state,
where they are invisible, and cannot interact with other objects.
The piece is kept in this state for a certain amount of time to
allow the client to catch up with potential RPC's that were sent,
in the event the client removed the piece before the server did.

For the host, this doesn't change anything, pieces are removed
from the scene tree directly and queued to be freed.