Remove extra entries, moved module classes to Global.
When a client joins a multiplayer game, any entries in the AssetDB
that the client has, that the host does not have, are temporarily
removed from the AssetDB until they leave the game.

Also moved both the TabletopImporter and ErrorReporter to Global,
which fixes a memory leak caused by the last commit.