Started sending chunks of missing asset files to the clients.
The maximum transfer speed is 500Kb/s. This is because of an
artificial delay in-between chunks, accounting for the fact that
the WebRTC buffer has a limited size, and we do not want to overflow
it.

If we had access to WebRTCDataChannel.get_buffered_amount(), we would
be able to check if there was outgoing data in the buffer and wait
accordingly, but we do not have access to that function (I think
because we're using the release version of the GDNative library).