Clients now save and import missing assets sent by the server.
However, I have had to remove the resource cache in PieceBuilder
to make this work (it kept saying there was a resource in cache when
there wasn't), so while the missing assets do get imported and they
can be spawned in, another issue has come up.

The reason I made the resource cache nearly a year ago in the commit
0166672e2bd8216572668501c865f1fbb220a761 was because there was a risk
of resource being freed and loaded at the same time - so I need to
find a way to prevent that from happening without a resource cache.

This commit fixes #11 (finally!), but I cannot merge the branch
yet due to the reasons stated above.