Fixed crash when closing the objects dialog.
I have FINALLY found the cause of the crash! It was caused by the
viewport trying to draw what was in it's world while the object
was being removed and freed.

The viewport now disables it's 3D functionality if it is not showing
anything. Once the disable command actually comes into effect (at the
end of the next render pass), only then is the child piece removed.