Clean collision shapes again, fixed error that came with cleaning.
I have something to admit... I... I was wrong. In commit
9471aa81aac71a35d29b5d53349ff3a9ca975178 I said that setting the
clean parameter to false when generating convex collision shapes
did not have an effect on the resulting shapes. I did some further
testing, and it turns out cleaning the shapes gives about a 10x
performance boost to the physics engine!

I have also fixed the error that came with the cleaning process -
it only happened when it tried to make a shape from a mesh with no
vertices in it. So the game no longer extracts a mesh instances with
no vertices in it.