Added the ResourceManager script.
It essentially just has load(), free(), and queue_free() but behind
a mutex so that these operations are thread-safe.

It also allows us to control if we want resources to stay in memory
after they are loaded for the first time - we do this on OSX for
StreamTextures, since I found a bug in the GLES3 implementation that
crashed the game when trying to re-draw a frame with a StreamTexture
that had just been freed.