ResourceManager now maintains its own free queue, clears on exit.
Previously, it queued objects to be freed via call_deferred, but this
was problematic, as sometimes the deferred calls were not processed
if the game exited abruptly (e.g. from an Alt+F4). Now, the objects
in the queue are guaranteed to be freed at the end of the program.