ObjectPreviewFactory is now a more generic PieceFactory.
When a script requests a piece be built by the PieceFactory, it is
given an "order number", which is used to accept the piece once it
has been built.

Both the ObjectPreview and FallingWorld scenes now use it - this fixes
a lot of stuttering when on the main menu on lower-end computers.