Aspect ratio is taken into account when drawing the fill-in line.
While trying to optimise the code, I came across a potential set of
methods belonging to the CanvasItem class, which can be used for custom
drawing - could these be used with a ViewportTexture, so that we don't
have to keep saving the image after the shader has been applied every
frame? Will test this in the future.