Created a new outline shader.
Instead of changing the albedo of a mesh depending on it's normal
vector at a particular pixel, a new mesh is now created within the
shader that is a slightly expanded version of the original mesh, with
front face culling as to not draw over the original.

This makes the highlighting of certain objects, e.g. cards, more
consistent.