Fixed the count label in the context menu putting cards in hands.
This was due to the label having the same ID, 0, as the first button
in the context menu, which happened to be the "put in hand" button.

All of the buttons in the context menu have had their IDs incremented
by 1, and all labels will have an ID of 0 by default.