Fixed stack content order desyncing between clients.
Sometimes, stacks of objects would not show the same order of items
for each of the clients, but taking one item off from the stack at
a time would always show the "correct" item for all clients.

This was caused by differing transforms between the clients for the
items and stacks - the transforms are used to determine at which end
of a stack an item is added to.

This has been fixed by adding extra arguments to the relevant RPCs
for the transforms, so the server can send it's transforms to all of
the clients - this should ensure that the calculations should remain
consistent across all peers without having to completely revamp the
netcode for stacks.