Added the ability to change how collision shapes are made.
By default, the game would create convex collision shapes for all 3D
models. While this works well for all of the pieces added to the game
so far, it will not suffice for more complicated pieces that I want
to have added later on, e.g. scenery or buildings.

Now, by using the 'collision_mode' property in the config.cfg file,
we can now specify whether we want a single convex collision shape,
multiple convex shapes, or a single concave shape. This will allow
the user to be able to choose how accurate they want the collision
shape to be, at the cost of the performance of the physics engine.

NOTE: Creating multiple convex shapes is quite computationally
expensive, so I am planning to only build pieces once, then store them
as a PackedScene in the user://assets directory, so the game does not
need to re-create the piece scene every time it is loaded.