Now create PackedScene caches of pieces that use collision_mode=1.
This was done to avoid an issue created by adding customisable collision
shapes: when creating multiple convex shapes (collision_mode = 1), the
game would freeze for up to a couple of seconds while creating the shapes,
which is annoying at best.

Now, for pieces that use this collision mode, two cached versions of the final
piece are saved as a PackedScene to the user://assets directory: one being
the in-game piece, and the other being the piece that is displayed in UI
previews. If these cached files are detected, and are not out of date, then
they are loaded and instanced instead of the game building the piece from
scratch again.

To help with debugging, in debug builds of the game a text version of the
PackedScene (.tscn) is created, whereas in a release build, a binary scene
is saved (.scn).

I had thought about using caches for all custom scene resources, but
in the end I decided it would be quite a lot of redundant file space
for not a whole lot of performance gain - when scenes are imported,
a .scn file is created in the user://.import directory, which in itself
contains a lot of what a PackedScene of the final piece would have.
Therefore, we would only really be saving the performance cost of
manipulating the scene tree, and creating extra nodes like CollisionShapes
and AudioEffectPlayers.