A minimum mouse speed is now required to fast-grab items.
This is to prevent micro-movements of the mouse from triggering a
fast grab, when the player intended to slow grab. Now, there is some
leeway for the mouse to move before a fast grab is triggered, so the
player does not need to keep the mouse perfectly still to slow grab.