Fixed objects changing albedo colour when copying and pasting.
This was due to the fact that if the object had a starting albedo
colour that was not pure white, then the state would have saved it
as a change in the colour, then Piece.set_albedo_color_client() would
have multiplied that value by the current albedo, hence making the
colour darker.

This was fixed by having Piece.get_albedo_color() report the current
albedo *relative* to the original colour. This way, even if the starting
albedo colour is not pure white, the state will not report a change
if it has not deviated from the original colour.