Added an extra check to a selected pieces loop to avoid a crash.
Very rarely, when deleting pieces, the "if piece is Piece:" line would
result in a crash, with the error being that the left-hand side was
already freed.

The inspector reports that the value of 'piece' when this occurs is
null, so I have added a null check in the loop to hopefully prevent
this crash from happening.