Adjusted the threshold of the audio compressor.
The threshold has been lowered to -6dB to try and prevent the audio
from getting too loud. This can be the case if, for example, a lot of
speakers are playing at the same time.

In the options menu, the slider-to-volume mapping has also been
adjusted. The maximum possible volume setting now maps to -6dB, and
the default master volume is now 0.5, which is equivalent to -12dB.