Added a pre-made save file for Poker to the default asset pack.
This save file however revealed a bug that has since been fixed:
when the state was loaded, and the player attempted to orient all of
the cards in the middle of the table downwards, the game would crash.

From what I could deduce, something goes wrong with the cache path
in the 'King of Clubs.png' and 'Back.png' resources (both of which
are initially loaded to help the stack form it's collision shape and
mesh instance), but deferring the call to re-generate the sandwich
stack stopped the crash from happening... somehow.