VSync is now enabled by	default.
Since the game is not that graphically demanding, modern computers
will not struggle to render the	game at	> 500 FPS, which is not	only
unnecessary, but could also cause 'coil	whining' on GPUs.

Therefore, I have decided to enable it by default if an options.cfg
file does not exist. Players with weaker machines always have the
option to disable it if their framerate is not reaching their monitor's
refresh rate.