Fixed an error creating config.cfg.md5 files under user://assets
The error would occur if the subdirectory was not yet created by
importing an asset, and an attempt to open a file at that location
was made.

This was fixed by checking if the directory existed before trying to
open the file - if it does not exist, it is created.