Servers now put pieces into limbo when in multiplayer.
This is in an effort to reduce the number of RPC errors that occur,
specifically when RPCs are sent from the client to the server.

The limbo system has also been tweaked to delete pieces after 3.0s
in the limbo group, rather than 5.0s after the last state update
(which did not account for hovered pieces on the server).

When in singleplayer, or alone in a multiplayer lobby, pieces are
deleted straight away as before.