Fixed pieces duplicating when popping undo states rapidly.
You know how when you find a bug and fix it, you're supposed to get
a feeling of joy, or gratification, like it's a reward? Well, this
wasn't it. Finding this hurt me to my core. I'm not the same man I
was before this. WHY DID I WRITE RETURN INSTEAD OF CONTINUE??!! ><