Fixed cards not being hidden entering hands through other hands.
This was due to the fact that cards could only keep track of one hand
at a time, so if there were two hands next to each other, then the
card would be seen as having left all hands when being hovered
through them both.

Now, cards keep an array of hands that they are currently inside,
so this should account for being in any number of hands at the same time.