Updated the README.
- Removed the Russian translation of the README. While I am incredibly
grateful for the contribution, any translations of the README are going
to keep going out of date with the main file. At some point in the future,
I will allow the README, as well as website content, to be translated via
Weblate, or a similar service.

- Removed the screenshot, as there are many on the itch.io page.

- Added a widget to the itch.io page - it will not be visible yet,
as the page is not public, but once it is it should automatically become
visible to everyone.