Added a workaround for reading the main assets folder on macOS.
FINALLY! You thought you could best me, Gatekeeper?? Of all of the
people you chose to mess with, you messed with me! ME!! You have NO
idea how much time I'm willing to spend on fixing something, and this
felt SO GOOD TO FIX. For ages macOS players would have to extract the
game in their Downloads folder so that the game could reach the default
asset pack from within quarantine, but once I learnt that Godot has
a function to return the executable path, and that .app bundles are just
folders... that was the moment the wall crumbled.

In case it was not immediately obvious...

I. AM. SO. HAPPY. :D