Can now change the master server URL with --master-server.
One should also use the new --ssl-certificate argument to override
the in-built certificate that identifies the official master server.

An example would be something like:
./TabletopClub.x86_64 --master-server wss://localhost:9080 --ssl-certificate public.crt

This should allow those who want to work on, or host their own master
server to connect to it without having to re-compile the game.

This commit closes #108.