Adjusted the way pieces are rotated when hovered.
Instead of changing the angular velocity to position the piece (which
causes issues when flipping pieces at weird angles, due to the variable
using euler angles), now the basis is set directly every physics frame
using Basis.slerp().

Also fixed spamming the flip button causing the basis to become the wrong
value. This was because the new hover_basis was based on the current
transform, not the current hover_basis.

This commit fixes #133.