Fixed containers not absorbing pieces when locked.
This was due to the fact that static bodies do not perform collision
detection, therefore the "body_entered" signal could never be emitted.

I had attempted to turn locked pieces into kinematic bodies, and while
this solved the issue, it created a lot of other ones (e.g. collisions
with rigid bodies are whack)

The solution I integrated was to override the lock methods for containers
on the server side, and instead of making them static bodies, simply just
locking the linear and angular axes, and disabling both in-built and
custom physics integration. This way, it acts the same way internally as
a static body, but it still performs collision detection.

This commit fixes #110.