Added a new directory to asset packs for boards.
This is to help distinguish game boards (e.g. the chess board currently
in the default asset pack) from game pieces, as they previously were
shown under the same type in the objects menu, "Pieces". As of now,
boards act in the same way as pieces in-game.

The Chess save game has been re-made to use the new board entry.

This commit closes #205.