Changed scale for cards to Vector2 (#210)
* Changed scale for cards to Vector2
Cards used to have a Vector3 for the scale property. However, a card is
scaled to 0.01 thickness by default. Increasing the thickness of cards
leads to a weird collider behavior. A card is thin and if an object is
thicker it should be of a different type (e.g. token).
To reflect that the thickness of cards can not be changed and for easier
configuration the scale was changed to Vector2.

* Force Vector3 for scale for all objects (except cards)

* Add warning if card scale is not Vector2

* Use Vector2 as default for card scale

* Set type for scale and default scale for assets

* Set scale to always be vector3 after config processing