Now clean old asset files before importing new ones.
This is to prevent files being left over after an asset pack is no
longer being imported (e.g. if the original folder was removed).

For example, since the chess board was under pieces/ and is now
under boards/, this update should remove the old files under pieces/
in the user:// directory.

This commit closes #232.