Can now configure centre-of-mass adjustment for 3D models.
This is done using the new 'com_adjust' property, which can take
one of three values: "off", "volume" (default, same functionality
as before), or "geometry".

This also solves the issue found in #142, in which d4 die would
bounce off of the table and settle on the wrong value when their
value was being set.

This commit closes #208.