Swapped scroll wheel behaviour when hovering pieces.
The default action is now to move the piece perpendicular to the
table, and holding down Ctrl now moves the piece towards and away
from the camera.

The key bindings for "Zoom In" and "Zoom Out" could already be used
to lift hovering pieces (since they just emulate the scroll wheel
anyways), and their entries in the options menu have been adjusted
to reflect this.

This commit closes #155.