Fixed objects from stacks hovering the wrong orientation on clients.
This commit is an extension of f42a826b7697a8ed97da8086afbbe6afe520b4f6,
which solved the issue for singleplayer and the host. Instead of having
a dedicated function for updating the '_last_physics_state_transform'
variable, it is now done automatically in start_hovering() if a physics
pass was not completed before the function was called.