Can now view other player's notebook pages if they are public.
A checkbox has been added to the notebook window for each page to
set whether it is public or not.

Clicking on a player's name below the room code will now show a context
menu, with the only option being (so far) to view that player's notebook.

The code for the notebook has also been completely re-written to be
more modular, in order to make adding more features in future commits
easier.