Added image-based page templates with configurable text boxes.
Images can now be used as the background of notebook pages, with
text boxes in the foreground that can be changed by the player.
This should allow creators to import things like character or score
sheets that players can use that will be more intuitive than a full
text file.

Currently, only the positions of the text boxes can be changed via
the config.cfg file, and the player cannot zoom in or out. These
will be addressed in future commits, this was mostly to lay the
foundation.

Also updated the AssetDB to potentially check the types of values
that are read from config files - this will stop the game freezing
if a value was not the type we expected.