Added a simple splash screen for the Godot Engine.
This commit is for #67, but I will keep the issue open for now
since the logo used is the one from Godot's press kit - in the future
it would be nice to have our own customised version of the logo!

The splash screen has a simple fade-in and fade-out effect, and can
be skipped either by pressing the "any" key, or via a new checkbox
in the options menu under "General".