Fixed errors when going through Object menu pages quickly.
One type of error that was occuring was the scene tree calling
`can_process()` on pieces that had just left the scene tree, but they
were still in the "idle_physics_process" group as the _physics_process
notification was being fired. This was solved by not removing the piece
from the scene tree early, and letting the queue_free() operation
remove it from the scene tree instead.

The game would also sometimes freeze at random intervals, with no errors
being thrown. After some debugging, I found that this freezing was
originating from the `load` call in `ResourceManager`, which was being
run inside a thread - thus I think the issue is the same as
godotengine/godot#55566. I implemented the same workaround mentioned
in that issue, and the game no longer freezes.

This commit closes #216.