Fixed positional audio button creating a callback loop.
This is similar to a previous issue where the range slider was
constantly changing value in multiplayer, which was solved by
detecting if the mouse was over the slider, and preventing server
RPCs from changing the value if that was the case.

Now, whenever the server sends an RPC, a flag is set to ignore signals
from the context menu controls temporarily until the RPC is resolved.