Stacks now preserve velocity when two pieces merge.
Previously, if two stackable pieces collided with a non-zero
velocity, the resulting stack would have no velocity. This could
be seen when placing the 52-card deck in a player's hand, and
throwing all the cards forward onto the table.

Now, the resulting stack will inherit the smallest linear and
angular velocities of the two pieces.

This commit fixes #194.