Now account for piece rotation when releasing from containers.
Previously, containers only took into account the piece height when
deciding where to place released pieces. This meant that if a piece
was rotated when entering the container, it may have immediately
collided with the container upon release, adding the piece back in.

Now, a new bounding box is calculated using the piece's rotation,
which is then used to decide how far away the piece should be from
the container upon release.