AssetDB entries are now resources instead of dictionaries.
This is the first major change in v0.2.0 - this change provides us
with type safety (no longer checking every second line if a value in
the dictionary is actually there), easy extendability (by making a
subclass of AssetEntry), and the ability to potentially include
custom assets inside the res:// directory, so the default asset pack
doesn't need to be imported when loading the game for the first time.
All of this for a little bit more memory usage, which is a good trade
in my opinion :P

Plus, in Godot 4.x, this change will actually lead to an increase in
performance, since the way GDScript handles types and properties of
objects has been optimised.

- Added various new data resources, e.g. CustomValue, DiceFaceValue.

- Added the SanityCheck class, which will provide us with... well,
sanity checks, throughout the project.

- Renamed RandomAudioStream to AudioStreamList.

- Added unit tests for both AssetEntry and SanityCheck.