Fixed stacks being added to stacks in the wrong position.
Previously, when a stack was being added, if a piece was flipped
internally within the stack (i.e. if the player had used "Orient Up"
or "Orient Down"), then it's entire transform was rotated around the
Z-axis. Crucially, this also included the origin, which is an error
because it meant that the other stack thought the cards being added
were under it, rather than over it.

Now, only the basis of the piece is flipped, rather than the whole
transform.