Added the new AssetPack class.
It stores data very similarly to before, that is, in multiple "type"
arrays in sorted order. However, the type arrays are now always
guaranteed to be present, it is a resource so packs can be created
in the editor, and it can handle adding and removing temporary
entries from the host.

- Added a new ResourceWithErrors class to store runtime errors for
multiple kinds of resources.