Added the TaggedDirectory class.
Files in a directory can now be tagged, which means metadata can be
stored about them which allows us to determine whether a file is new,
or has changed since it was last tagged.

Untagged files can also be removed from the directory all at once,
which allows us to remove rogue files from the user:// directory.

This class will be the foundation for future importing classes.