Started work on configuring asset entries from .cfg files.
Instead of entries being created in one giant function, this time
around entries will be created in two stages: the first being setup,
which will be done by AssetPackCatalog, and is mostly determined by
the directory the asset resides in. The second is configuring, which
applies the values from the .cfg file on top of what is already in
the entry after the setup stage. This, on top of the in-built data
validation within the entry classes, should make the code a lot more
readable, while also making it generic enough to potentially use it
elsewhere if need be.

To help with this, the DictionaryParser class was created, which
serves a similar function to AdvancedConfigFile in that it reads
values from dictionaries with extra data validation.