Finished code for configuring asset entries.
It should now be much easier and safer to introduce new changes to
how the 'config.cfg' file is read. Now, I just need to have an easy
way to set up scene entries, to read 'stacks.cfg' files to make
collections of entries, and to test everything as well... ho boy ><
It will be worth it in the end, but this is what I dreaded most
about using unit tests, some of them are going to be absolutely
gargantuan, and I'm pretty sure the tests for this code is going
to be one of them... :|

Overall, the functionality has mostly stayed the same from before,
except for one major change: the 'face_values' dictionary used for
dice has swapped its keys and values. Previously, the keys were the
values assigned to each face, and the values were the rotations needed
to make that face point globally upwards. Now, the keys are the
rotations, and the values are the, well, values. This allows for
multiple faces to have not only the same value, but string values
are also now possible due to the CustomValue class.

As a result, once this commit is merged into master, it should
close #209.